+++ Editor/Content/ContentBrowser/PlainView.cs
@@ -1,0 +1,46 @@
+﻿using System.Windows;
+using System.Windows.Controls;
+
+namespace Editor.Content.ContentBrowser
+{
+    class PlainView : ViewBase
+    {
+        protected override object DefaultStyleKey => new ComponentResourceKey(GetType(), "PlainViewResourceId");
+
+        public static readonly DependencyProperty ItemContainerStyleProperty =
+            ItemsControl.ItemContainerStyleProperty.AddOwner(typeof(PlainView));
+
+        public static readonly DependencyProperty ItemTemplateProperty =
+            ItemsControl.ItemTemplateProperty.AddOwner(typeof(PlainView));
+
+        public static readonly DependencyProperty ItemWidthProperty =
+            WrapPanel.ItemWidthProperty.AddOwner(typeof(PlainView));
+
+        public static readonly DependencyProperty ItemHeightProperty =
+            WrapPanel.ItemHeightProperty.AddOwner(typeof(PlainView));
+
+        public Style ItemContainerStyle
+        {
+            get => (Style)GetValue(ItemContainerStyleProperty);
+            set => SetValue(ItemContainerStyleProperty, value);
+        }
+
+        public DataTemplate ItemTemplate
+        {
+            get => (DataTemplate)GetValue(ItemTemplateProperty);
+            set => SetValue(ItemTemplateProperty, value);
+        }
+
+        public double ItemWidth
+        {
+            get => (double)GetValue(ItemWidthProperty);
+            set => SetValue(ItemWidthProperty, value);
+        }
+
+        public double ItemHeight
+        {
+            get => (double)GetValue(ItemHeightProperty);
+            set => SetValue(ItemHeightProperty, value);
+        }
+    }
+}