+++ Editor/Utilities/Converters/DataSizeToStringConverter.cs
@@ -1,0 +1,49 @@
+﻿using System.Globalization;
+using System.Windows.Data;
+
+namespace Editor.Utilities.Converters
+{
+    public class DataSizeToStringConverter : IValueConverter
+    {
+        private static readonly string[] _sizeSuffixes =
+        {
+            "B",
+            "kB",
+            "MG",
+            "GB",
+            "TB",
+            "PB",
+            "EB",
+            "ZB",
+            "YB"
+        };
+
+        static string SizeSuffix(long value, int decimalPlaces = 1)
+        {
+            if (value <= 0 || decimalPlaces < 0) return string.Empty;
+
+            // mag is 0 for bytes, 1 for kB, 2 for MB, etc.
+            int mag = (int)Math.Log(value, 1024);
+
+            // 1L << (mag * 10) == 2 ^ (10 * mag)
+            // i.e. the number of bytes in the unit corresponding to mag
+            decimal adjustedSize = (decimal)value / (1L << (mag * 10));
+
+            if (Math.Round(adjustedSize, decimalPlaces) >= 1000)
+            {
+                mag += 1;
+                adjustedSize /= 1024;
+            }
+
+            return string.Format("{0:n" + decimalPlaces + "} {1}", adjustedSize, _sizeSuffixes[mag]);
+        }
+
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            return (value is long size) ? SizeSuffix(size, 0) : null;
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture) =>
+            throw new NotImplementedException();
+    }
+}